chore: restructure project to convention
Move the library code to a new dir, wiktionaryparser, so the project
conforms with the structure of other Python projects, but also so
absolute imports and intellisense works, given the dev ran

    install -e .

Also added a CONTRIBUTING.md with specific instructions that
contributors can copy-paste to their bash terminal to quickly setup
their dev environment.

Finally, added pylint to the requirements.txt.