Fix wrong language issue for words with no contents
If a word has no contents box which would be used to find that word's contents (such as "1. English" or "1.2.1.2 Related terms") in page, the language given as parameter wouldn't be checked for equality.

This means that even if you want to get only the definitions for English words, all definitions would be returned, regardless of their language.

For example, the word "cartier" will return its French definition, even when you give the English language as parameter (or any other language, for that matter).

This commit makes it so that the output for fetch("cartier", "English"), instead of:

[{'definitions': [{'examples': [], 'partOfSpeech': 'noun', 'relatedWords': [], 'text': ['cartier\xa0m (plural cartiers)', 'playing card maker']}],
  'etymology': '',
  'pronunciations': {'audio': [], 'text': []}}]

 becomes:

 []

 while fetch("cartier", "French") still returns that definition.