fix: changed return structure to pass test
A return_categories option is added to the fetch
function defaulting to false; with this option set
to false, the fetch function will return the
original word information in json format.
If this option is set to true, the function will
return a pair of the word information and a list
of its categories. This change was made to make sure
the function passes the unit test.